Stressing name resolution for deprecated and aliased backends
